+++ src/pages/index.js
@@ -5,10 +5,10 @@ import { pageAnimation } from '../animation/animation'
 export default function Home() {
   return (
     <motion.div
-      exit='exit'
+      exit="exit"
       variants={pageAnimation}
-      initial='hidden'
-      animate='show'
+      initial="hidden"
+      animate="show"
     >
       <Hero showHero={true} />
       <Experiences />